Cannot find method handle with parameter types: [org.eclipse.jetty.server.HttpChannel] #1257
support jetty 8.x version.
